backend/requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
sqlalchemy==2.0.20
psycopg2-binary==2.9.6
pydantic==2.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==4.6.0
stripe==7.0.0
openai==0.27.8
python-dotenv==1.0.0
sendgrid==6.10.0
alembic==1.11.1
pytest==7.4.0
pytest-asyncio==0.21.1
httpx==0.24.1